Renaming functions, renaming the structure, updating messages and improving the use of the application.
