+++ App/HiveHub/CMakeLists.txt
@@ -10,8 +10,8 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 # Add source to the executable file of this project.
-file(GLOB_RECURSE SOURCES "main.cpp" "Config/*.cpp" "Database/*.cpp" "Methods/*.cpp" "Methods/Edits_Methods/*.cpp" "Methods/Display_Methods/*.cpp" "Methods/GUI_Methods/*.cpp")
-file(GLOB_RECURSE HEADERS "*.h"  "Config/*.h" "Database/*.h" "Methods/*.h" "Methods/Display_Methods/*.h" "Methods/GUI_Methods/*.h")
+file(GLOB_RECURSE SOURCES "main.cpp" "Config/*.cpp" "Database/*.cpp" "Methods/*.cpp" "Methods/Edits_Features/*.cpp" "Methods/User_Display_Features/*.cpp" "Methods/GUI_Features/*.cpp")
+file(GLOB_RECURSE HEADERS "*.h"  "Config/*.h" "Database/*.h" "Methods/*.h" "Methods/User_Display_Features/*.h" "Methods/GUI_Features/*.h")
 
 add_executable(HiveHub ${SOURCES} ${HEADERS})
