+++ App/HiveHub/Methods/User_Display_Features/display_all_users.cpp
@@ -1,4 +1,4 @@
-#include "display_methods.h"
+#include "display_features.h"
 
 void displayEveryone(WriteDatabase& writeDatabase) {
 	cout << "[MAIN|DISPLAY|ALL USERS]" << endl << endl;
