+++ App/HiveHub/Methods/miscellaneous_functions.cpp
