+++ App/HiveHub/Methods/person_addition.cpp
@@ -2,8 +2,8 @@
 
 void userCreator(WriteDatabase& writeDatabase) {
 	cout << "[MAIN|CREATION]" << endl << endl;
-	cout << "Welcome to the user Creation Tool. This tool allows you to add a new person to the database." << endl;
-	cout << "Please prepare data in the following format: [Email | FirstName | LastName | Address | Interests]" << endl << endl;
+	cout << "Welcome to the User Creation Tool. This tool allows you to add a new person to the database." << endl;
+	cout << "Please enter the required information in the following format: [Email | FirstName | LastName | Address | Interests]" << endl << endl;
 
 	string email, firstName, lastName, address, interestInput;
 	vector<string> interests;
@@ -29,7 +29,7 @@ void userCreator(WriteDatabase& writeDatabase) {
 	cout << "Enter Address: ";
 	getline(cin, address);
 
-	cout << "Enter Interests (comma-separated without space): ";
+	cout << "Enter Interests (comma-separated without spaces): ";
 	getline(cin, interestInput);
 
 	istringstream iss(interestInput);
@@ -43,14 +43,14 @@ void userCreator(WriteDatabase& writeDatabase) {
 	ReadDatabase readDatabase(writeDatabase);
 	if (readDatabase.findUserByEmail(email)) {
 		cout << endl << "####################" << endl;
-		cout << "ERROR:. This email already exists in the database! Please provide an email that doesn't exist." << endl;
+		cout << "Error: This email already exists in the database! Please provide a unique email." << endl;
 		cout << "####################" << endl << endl;
 		system("PAUSE");
 	}
 	else {
 		writeDatabase.addUser(email, firstName, lastName, address, interests);
 		cout << endl << "####################" << endl;
-		cout << "SUCCESS:. User added to the database successfully." << endl;
+		cout << "Success: User added to the database successfully." << endl;
 		cout << "####################" << endl << endl;
 		system("PAUSE");
 	}
