+++ App/HiveHub/main.h
@@ -2,7 +2,7 @@
 
 #include "common_includes.h"
 
-#include "Methods/GUI_Methods/user_interface.h"
+#include "Methods/GUI_Features/user_interface.h"
 #include "Database/database_controller.h"
 #include "Config/config.h"
 #include "Config/console_utils.h"
